Simplify component
